Separated update school/admin command logic and admin call into (update_academics).
